refactor: menugroup 패키지 분리
