+++ src/main/java/kitchenpos/domain/menugroup/MenuGroupRepository.java
@@ -1,6 +1,5 @@
-package kitchenpos.dao;
+package kitchenpos.domain.menugroup;
 
-import kitchenpos.domain.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
