+++ src/main/java/kitchenpos/ui/menugroup/MenuGroupRestController.java
@@ -1,10 +1,10 @@
-package kitchenpos.ui;
+package kitchenpos.ui.menugroup;
 
 import java.net.URI;
 import java.util.List;
-import kitchenpos.application.MenuGroupService;
-import kitchenpos.dto.MenuGroupRequest;
-import kitchenpos.dto.MenuGroupResponse;
+import kitchenpos.application.menugroup.MenuGroupService;
+import kitchenpos.dto.menugroup.MenuGroupRequest;
+import kitchenpos.dto.menugroup.MenuGroupResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
